📍 Rename Hospitals path to Hospital
